Project Commit
This project was done following the YouTube video tutorial uploaded by Avi Singh. The pdf attached shows the dashboard created at the end of the course. I went through this course as an introduction to Power BI, in the span of 1 hour which I have found to be rather useful.

Disclaimer : The source code that was used to implement the coursework project is no longer maintained. There may be errors or bugs that did not exist at the time of creation.